backend/requirements.txt
astroid==2.15.0
cffi==1.15.1
click==8.1.3
colorama==0.4.6
configobj==5.0.8
cryptography==40.0.2
dill==0.3.6
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
future==0.18.3
greenlet==2.0.0
gunicorn==20.1.0
importlib-metadata==5.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.1
mccabe==0.7.0
platformdirs==3.1.1
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pylint==2.17.0
python-dotenv==0.21.0
python-gnupg==0.5.0
python-ldap==3.4.3
secrets==1.0.2
six==1.16.0
SQLAlchemy==1.4.42
systematic==4.8.7
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
Werkzeug==2.2.2
wrapt==1.15.0
zipp==3.10.0
